+++ src/score/pub/wms-controls.ts
@@ -144,18 +144,6 @@ export class WmsControls {
         return this;
     }
 
-    /**
-     * Set play button.
-     * @deprecated setPlayButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.
-     * @param playBtn - HTML button element or element id.
-     * @param playLabel - Custom button label (e.g. "Play").
-     * @returns - This playback buttons class instance.
-     */
-    setPlayButton(playBtn: HTMLButtonElement | string, playLabel?: string): WmsControls {
-        warnOnce("setPlayButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.");
-        return this._setPlayButton(playBtn, playLabel);
-    }
-
     private _setStopButton(stopBtn: HTMLButtonElement | string, stopLabel?: string): WmsControls {
         AssertUtil.assertVar(Guard.isStringOrUndefined(stopLabel), "stopLabel", stopLabel);
 
@@ -170,18 +158,6 @@ export class WmsControls {
         this.updateButtons();
 
         return this;
-    }
-
-    /**
-     * Set stop button.
-     * @deprecated setStopButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.
-     * @param stopBtn - HTML button element or element id.
-     * @param stopLabel - Custom button label (e.g. "Stop").
-     * @returns - This playback buttons class instance.
-     */
-    setStopButton(stopBtn: HTMLButtonElement | string, stopLabel?: string): WmsControls {
-        warnOnce("setStopButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.");
-        return this._setStopButton(stopBtn, stopLabel);
     }
 
     private _setPlayStopButton(playStopBtn: HTMLButtonElement | string, playLabel?: string, stopLabel?: string): WmsControls {
@@ -202,19 +178,6 @@ export class WmsControls {
         return this;
     }
 
-    /**
-     * Set play/stop button.
-     * @deprecated setPlayStopButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.
-     * @param playStopBtn - HTML button element or element id.
-     * @param playLabel - Custom button label for play action (e.g. "Play").
-     * @param stopLabel - Custom button label for stop action (e.g. "Stop").
-     * @returns - This playback buttons class instance.
-     */
-    setPlayStopButton(playStopBtn: HTMLButtonElement | string, playLabel?: string, stopLabel?: string): WmsControls {
-        warnOnce("setPlayStopButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.");
-        return this._setPlayStopButton(playStopBtn, playLabel, stopLabel);
-    }
-
     private _setPauseButton(pauseBtn: HTMLButtonElement | string, pauseLabel?: string): WmsControls {
         AssertUtil.assertVar(Guard.isStringOrUndefined(pauseLabel), "pauseLabel", pauseLabel);
 
@@ -229,18 +192,6 @@ export class WmsControls {
         this.updateButtons();
 
         return this;
-    }
-
-    /**
-     * Set pause button.
-     * @deprecated setPauseButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.
-     * @param pauseBtn - HTML button element or element id.
-     * @param pauseLabel - Custom button label (e.g. "Pause").
-     * @returns - This playback buttons class instance.
-     */
-    setPauseButton(pauseBtn: HTMLButtonElement | string, pauseLabel?: string): WmsControls {
-        warnOnce("setPauseButton() is deprecated. Will be removed in future release. Use `setSinglePlay()`, `setSinglePlayStop()`, `setPlayStop()` or `setPlayPauseStop()` instead.");
-        return this._setPauseButton(pauseBtn, pauseLabel);
     }
 
     private static savedOnClickListeners = new UniMap<HTMLButtonElement, (() => void)[]>();